# solve for ne instead of using atmospheric value