--- a/src/statmech.jl
+++ b/src/statmech.jl
@@ -99,7 +99,6 @@ function chemical_equilibrium(T, nₜ, model_atm_nₑ, absolute_abundances, ioni
                               partition_fns, log_equilibrium_constants; x0=nothing)
     if x0 === nothing
         x_type = promote_type(eltype(absolute_abundances), typeof(model_atm_nₑ), typeof(T), typeof(nₜ))
-        println(x_type)
         x0 = Vector{x_type}(undef, MAX_ATOMIC_NUMBER + 1)
         #compute good first guess by neglecting molecules
         for Z in 1:MAX_ATOMIC_NUMBER
@@ -112,7 +111,7 @@ function chemical_equilibrium(T, nₜ, model_atm_nₑ, absolute_abundances, ioni
     #numerically solve for equlibrium.
     residuals! = chemical_equilibrium_equations(T, nₜ, absolute_abundances, ionization_energies, 
                                                 partition_fns, log_equilibrium_constants)
-    sol = nlsolve(residuals!, x0; iterations=1_000, store_trace=true, ftol=1e-8, autodiff=:forward)
+    sol = nlsolve(residuals!, x0; method=:newton, iterations=1_000, store_trace=true, ftol=1e-8, autodiff=:forward)
     if !sol.f_converged
         error("Molecular equlibrium unconverged. \n", sol)
     elseif !all(isfinite, sol.zero)