This branch is an experiment where Korg solves for the electron number density self-consistently, rather than assuming the model-atmospheric value is correct.  For the MARCS solar atmosphere, this adds about 0.1s to synthesis time, which is roughly double for a small wavelength range.

Notably, this implementation doesn't crash on any model atmospheres (at least at solar metallicity).  Contrast with the current method (#129), which sometimes fails for cool stars.

Here's the reletive difference in the $n_e$ that Korg infers using the old method and this one.  Solar atmosphere on the left, cool star (Teff = 3200K, logg=2.0 very close to the parameters which crash the current implementation) on the right:
![image](https://user-images.githubusercontent.com/711963/235231020-d6ca96f6-a569-43ca-af51-7a7c701d76d0.png)![image](https://user-images.githubusercontent.com/711963/235230887-f71cbcea-ed73-41af-874d-b46bc713b1d2.png)

The agreement between the methods is worse for the cool star, which is consistent with Korg's solver failing near there.  Completely unexpectedly (to me), the new method actually produces closer agreement with the model atmosphere pretty much everywhere. I'd really like to know what the cause for disagreement with the atmosphere is, but my guess is charged molecules.

The seems like a slam dunk except that it's slow.  

TODO 
- [x] actually use new ne for continuum stuff, etc
- [x] make fast?
- [x] document changes in atmosphere code
- [x] add warning when electron number density diverges from model atmosphere
- [x] stress test to see where, if anywhere, Korg's chemical equilibrium can't converge