@ajwheeler 2023-05-10
OK, so I spent some time speeding this up and it's actually a bit faster than the old implementation now for cool stars where chemical equilibrium is more complicated (more molecules). 

new implementation: 70ms (sun),  148ms (cool giant)
old implementation: 70ms (sun) 227 ms (cool giant)